src/main/java/com/taller/core/dominio/Backlog.java
package com.taller.core.dominio;

import java.util.List;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.OneToMany;
import javax.persistence.OneToOne;
import javax.validation.constraints.NotBlank;

import com.fasterxml.jackson.annotation.JsonIgnore;

@Entity
public class Backlog {

	@Id
	private Long id;
	
	@NotBlank (message = "El atributo projectIdentifier no puede estar vacío")
	private String projectIdentifier;
	
	@JsonIgnore
	@OneToOne(mappedBy="backlog")
	private Project project;
	
	@JsonIgnore
	@OneToMany(mappedBy="backlog")
	private List<ProjectTask> projectTasks;
	
	
	public Long getId() {
		return id;
	}
	public void setId(Long id) {
		this.id = id;
	}
	public String getProjectIdentifier() {
		return projectIdentifier;
	}
	public void setProjectIdentifier(String projectIdentifier) {
		this.projectIdentifier = projectIdentifier;
	}
	public Project getProject() {
		return project;
	}
	public void setProject(Project project) {
		this.project = project;
	}
	public List<ProjectTask> getProjectTasks() {
		return projectTasks;
	}
	public void setProjectTasks(List<ProjectTask> projectTasks) {
		this.projectTasks = projectTasks;
	}
	
}
